Fix binding of `handleInputClick`
Had missed that is was not bound like the other `handleInput...` functions
when removing the wrapping arrow function in 25cc4a3.

This was displaying an error in the browser console as `this` was not defined
when the method was called in response to clicks.

That error in the console seems the only effect of the function not being bound.
Both with and without input, and with or without the `autoselect` option set,
I did not notice any visual change in behaviour from the component.

We may want to review the purpose of that event handler, maybe the `handleInputFocus` one does enough? 🤔